--- src/main/java/com/grepp/nbe562team04/model/achieve/rule/TutorialAchievement.java
@@ -1,4 +1,0 @@
-package com.grepp.nbe562team04.model.achieve.rule;
-
-public class TutorialAchievement {
-}
