+++ src/main/java/com/grepp/nbe562team04/model/achievement/rule/TutorialAchievement.java
@@ -1,0 +1,4 @@
+package com.grepp.nbe562team04.model.achievement.rule;
+
+public class TutorialAchievement {
+}
