+++ src/main/java/com/grepp/nbe562team04/model/goal/GoalService.java
@@ -1,5 +1,7 @@
 package com.grepp.nbe562team04.model.goal;
 
+import com.grepp.nbe562team04.model.achievement.AchievementService;
+import com.grepp.nbe562team04.model.achievement.entity.Achievement;
 import com.grepp.nbe562team04.model.goal.dto.GoalRequestDto;
 import com.grepp.nbe562team04.model.goal.dto.GoalResponseDto;
 import com.grepp.nbe562team04.model.goal.entity.Goal;
@@ -25,10 +27,10 @@ public class GoalService {
     private final GoalCompanyRepository goalCompanyRepository;
     private final TodoRepository todoRepository;
     private final UserRepository userRepository;
-
+    private final AchievementService achievementService;
     // 목표 생성
     @Transactional
-    public void createGoal(GoalRequestDto dto) {
+    public String createGoal(GoalRequestDto dto, Long userId) {
         GoalCompany company = goalCompanyRepository.findById(dto.getCompanyId())
                 .orElseThrow(() -> new RuntimeException("해당 기업이 존재하지 않습니다."));
 
@@ -42,6 +44,7 @@ public class GoalService {
                 .build();
 
         goalRepository.save(goal);
+        return achievementService.giveFirstGoalCreateAchievement(userId);
     }
 
     // 기업별 목표 목록 조회
@@ -138,5 +141,15 @@ public class GoalService {
 
         user.addXp(10);
         userRepository.save(user);
+    }
+
+    public String giveFirstGoalAchievementIfNeeded(User user) {
+        Long userId = user.getUserId();
+        long count = goalRepository.countByCompany_User_UserId(userId);
+
+        if (count == 1) { // 처음 생성한 경우
+            return achievementService.giveFirstGoalCreateAchievement(userId); // 이름 반환
+        }
+        return null;
     }
 }
